Add metadata to pgaudit log logline
Previously we were using project-id/endpoint-id as SYSLOGTAG, which has a
limit of 32 characters, so the endpoint-id got truncated. Instead,
prepend the metadata as key=value pairs (logfmt) to the log line, which
is easy enough to parse, without having to deal with SYSLOG tag length
limitation.

Additionally, the configuration forwarded just everything, but we are
interested in pgaudit specifically. Using local5 facility for that,
which is currenly unused.